Update to Ratatui 0.21
Cargo.toml is updated to specify 0.21 rather than 0.* as 0.21 deprecates
the `Spans` type and replaces it with `Line`. This is generally a good
thing as it makes the code more readable and easier to use.

A bunch of tests are updated to use the new `Line` type and are
simplified to use the Text::raw / Text::styled functions rather than
constructing the `Spans` type directly.